pyproject.toml
[tool.poetry.scripts]
qc = "azconnect.app:app"

[tool.poetry]
name = "azconnect"
version = "0.1.0"
description = "A cli tool for connecting to Kubernetes Clusters"
authors = ["Joe MacMillan <joe@quasar.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
azure-cli = "^2.64.0"
typer = "^0.12.5"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
